# services/server/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Binary Brigade",
        "version": "1.0.0",
        "title": "Patient Tracker application",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
...
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
        "schemas": {
...
            "user_create-profile": {
                "type": "object",
                "properties": {
                    "email_id": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "user_type": {
                        "type": "integer"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "address": {
                        "type": "string"
                    },
                    "contact_no": {
                        "type": "string"
                    },
                    "dob": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer"
                    },
                    "gender": {
                        "type": "integer"
                    },
                    "qualifications": {
                        "type": "string"
                    },
                    "keywords": {
                        "type": "integer"
                    },
                    "last_login_date": {
                        "type": "integer"
                    },
                    "registration_date": {
                        "type": "integer"
                    },
                    "update_date": {
                        "type": "integer"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/api/user/create-profile": {
            "post": {
                "tags": [
                    "Create profile"
                ],
                "summary": "Create profile",
                "requestBody": {
                    "description": "Create User profile in Database for doctor and patient",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user_create-profile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/user_create-profile"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/user_create-profile"
                        }
                    }
                }
            }
        },
...
    }
}